backend/requirements-flask.txt
Flask==3.1.0
onnxruntime==1.20.1
numpy
fastapi
sqlmodel
PyJWT  # Remove if not using JWT authentication
uvicorn
pydantic
opencv-python-headless==4.11.0.86
insightface==0.7.3  # Remove if not using face recognition
huggingface-hub==0.29.2  # Remove if not loading models from Hugging Face
